+++ src/routes/register/+page.server.ts
@@ -12,9 +12,6 @@ export const actions = {
 		const formData = await request.formData();
 		const email = formData.get('email') as string;
 		const password = formData.get('password') as string;
-		if (password.length < 8) {
-			return 'Password must be at least 8 characters long.';
-		}
 		const session = await trpc(locals.auth).users.register({ email, password });
 		if (session !== null) {
 			locals.auth.setSession(session);